fastapi-server/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==4.2.4
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==1.7.3
google-cloud-speech==2.27.0
google-cloud-translate==2.0.1
google-cloud-vision==3.7.4
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.10
requests==2.32.3
rich==13.7.1
rsa==4.9
scipy==1.14.0
seaborn==0.13.2
setuptools==70.1.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
starlette==0.37.2
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.48
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
